rearrange radial axis stories
