+++ src/RadialAreaChart/RadialAreaChart.story.tsx
@@ -100,277 +100,3 @@ export const MultiSeries = () => (
     axis={<RadialAxis type="category" />}
   />
 );
-
-export const SemiCircle = () => (
-  <RadialAreaChart
-    height={500}
-    width={500}
-    data={medDateData}
-    innerRadius={0.1}
-    series={
-      <RadialAreaSeries
-        colorScheme="cybertron"
-        animated
-        interpolation="smooth"
-        area={<RadialArea gradient={<RadialGradient />} />}
-      />
-    }
-    axis={
-      <RadialAxis
-        arcs={<RadialAxisArcSeries count={10} />}
-        ticks={
-          <RadialAxisTickSeries
-            count={5}
-            tick={
-              <RadialAxisTick
-                line={<RadialAxisTickLine position="inside" />}
-                label={<RadialAxisTickLabel autoRotate />}
-              />
-            }
-          />
-        }
-      />
-    }
-    startAngle={-0.5 * Math.PI}
-    endAngle={0.5 * Math.PI}
-  />
-);
-
-export const SemiCircleMultiSeries = () => (
-  <RadialAreaChart
-    data={multiCategory}
-    height={500}
-    width={500}
-    series={
-      <RadialAreaSeries
-        type="grouped"
-        interpolation="linear"
-        area={
-          <RadialArea
-            gradient={
-              <RadialGradient
-                stops={[
-                  <GradientStop offset="0%" stopOpacity={0.1} key="start" />,
-                  <GradientStop offset="80%" stopOpacity={0.3} key="stop" />
-                ]}
-              />
-            }
-          />
-        }
-      />
-    }
-    axis={<RadialAxis type="category" />}
-    startAngle={-0.5 * Math.PI}
-    endAngle={0.5 * Math.PI}
-  />
-);
-
-export const QuarterCircle = () => (
-  <RadialAreaChart
-    height={500}
-    width={500}
-    data={medDateData}
-    innerRadius={0.1}
-    series={
-      <RadialAreaSeries
-        colorScheme="cybertron"
-        animated
-        interpolation="linear"
-        area={<RadialArea gradient={<RadialGradient />} />}
-      />
-    }
-    axis={
-      <RadialAxis
-        arcs={<RadialAxisArcSeries count={10} />}
-        ticks={
-          <RadialAxisTickSeries
-            count={5}
-            tick={
-              <RadialAxisTick
-                line={<RadialAxisTickLine position="inside" />}
-                label={<RadialAxisTickLabel autoRotate />}
-              />
-            }
-          />
-        }
-      />
-    }
-    startAngle={0}
-    endAngle={0.5 * Math.PI}
-    isClosedCurve={false}
-  />
-);
-
-export const QuarterCircleMultiSeries = () => (
-  <RadialAreaChart
-    data={multiCategory}
-    height={500}
-    width={500}
-    series={
-      <RadialAreaSeries
-        type="grouped"
-        interpolation="linear"
-        area={
-          <RadialArea
-            gradient={
-              <RadialGradient
-                stops={[
-                  <GradientStop offset="0%" stopOpacity={0.1} key="start" />,
-                  <GradientStop offset="80%" stopOpacity={0.3} key="stop" />
-                ]}
-              />
-            }
-          />
-        }
-      />
-    }
-    axis={<RadialAxis type="category" />}
-    startAngle={0}
-    endAngle={0.5 * Math.PI}
-    isClosedCurve={false}
-  />
-);
-
-export const CustomCircle = () => (
-  <RadialAreaChart
-    height={500}
-    width={500}
-    data={medDateData}
-    innerRadius={0.1}
-    series={
-      <RadialAreaSeries
-        colorScheme="cybertron"
-        animated
-        interpolation="linear"
-        area={<RadialArea gradient={<RadialGradient />} />}
-      />
-    }
-    axis={
-      <RadialAxis
-        arcs={<RadialAxisArcSeries count={10} />}
-        ticks={
-          <RadialAxisTickSeries
-            count={5}
-            tick={
-              <RadialAxisTick
-                line={<RadialAxisTickLine position="inside" />}
-                label={<RadialAxisTickLabel autoRotate />}
-              />
-            }
-          />
-        }
-      />
-    }
-    startAngle={-Math.PI}
-    endAngle={0.5 * Math.PI}
-    isClosedCurve={false}
-  />
-);
-
-export const CustomCircle2 = () => (
-  <RadialAreaChart
-    height={500}
-    width={500}
-    data={medDateData}
-    innerRadius={0.1}
-    series={
-      <RadialAreaSeries
-        colorScheme="cybertron"
-        animated
-        interpolation="linear"
-        area={<RadialArea gradient={<RadialGradient />} />}
-      />
-    }
-    axis={
-      <RadialAxis
-        arcs={<RadialAxisArcSeries count={10} />}
-        ticks={
-          <RadialAxisTickSeries
-            count={5}
-            tick={
-              <RadialAxisTick
-                line={<RadialAxisTickLine position="inside" />}
-                label={<RadialAxisTickLabel autoRotate />}
-              />
-            }
-          />
-        }
-      />
-    }
-    startAngle={-Math.PI}
-    endAngle={-0.25 * Math.PI}
-    isClosedCurve={false}
-  />
-);
-
-export const CustomCircle3 = () => (
-  <RadialAreaChart
-    height={500}
-    width={500}
-    data={medDateData}
-    innerRadius={0.1}
-    series={
-      <RadialAreaSeries
-        colorScheme="cybertron"
-        animated
-        interpolation="smooth"
-        area={<RadialArea gradient={<RadialGradient />} />}
-      />
-    }
-    axis={
-      <RadialAxis
-        arcs={<RadialAxisArcSeries count={10} />}
-        ticks={
-          <RadialAxisTickSeries
-            count={5}
-            tick={
-              <RadialAxisTick
-                line={<RadialAxisTickLine position="inside" />}
-                label={<RadialAxisTickLabel autoRotate />}
-              />
-            }
-          />
-        }
-      />
-    }
-    startAngle={-0.75 * Math.PI}
-    endAngle={0.75 * Math.PI}
-    isClosedCurve={false}
-  />
-);
-
-export const CustomCircle4 = () => (
-  <RadialAreaChart
-    height={500}
-    width={500}
-    data={medDateData}
-    innerRadius={0.1}
-    series={
-      <RadialAreaSeries
-        colorScheme="cybertron"
-        animated
-        interpolation="linear"
-        area={<RadialArea gradient={<RadialGradient />} />}
-      />
-    }
-    axis={
-      <RadialAxis
-        arcs={<RadialAxisArcSeries count={10} />}
-        ticks={
-          <RadialAxisTickSeries
-            count={5}
-            tick={
-              <RadialAxisTick
-                line={<RadialAxisTickLine position="inside" />}
-                label={<RadialAxisTickLabel autoRotate />}
-              />
-            }
-          />
-        }
-      />
-    }
-    startAngle={-0.25 * Math.PI}
-    endAngle={0.25 * Math.PI}
-    isClosedCurve={false}
-  />
-);
