+++ src/RadialAreaChart/RadialAxisLineChart.story.tsx
@@ -1,4 +1,4 @@
-import { categoryData, medDateData, multiCategory } from '../../demo';
+import { medDateData, multiCategory } from '../../demo';
 import {
   RadialAxis,
   RadialAxisArcSeries,
@@ -16,7 +16,7 @@ import {
 } from './RadialAreaSeries';
 
 export default {
-  title: 'Charts/Line Chart/Radial',
+  title: 'Charts/Line Chart/Radial/Axis',
   component: RadialAreaChart,
   subcomponents: {
     RadialAreaSeries,
@@ -26,7 +26,7 @@ export default {
   }
 };
 
-export const Simple = () => (
+export const SemiCircle = () => (
   <RadialAreaChart
     height={500}
     width={500}
@@ -56,25 +56,55 @@ export const Simple = () => (
         }
       />
     }
+    startAngle={-0.5 * Math.PI}
+    endAngle={0.5 * Math.PI}
   />
 );
 
-export const CategoricalData = () => (
-  <RadialAreaChart
-    data={categoryData}
-    height={500}
-    width={500}
-    series={<RadialAreaSeries area={null} />}
-    axis={<RadialAxis type="category" />}
-  />
-);
-
-export const MultiSeries = () => (
+export const SemiCircleMultiSeries = () => (
   <RadialAreaChart
     data={multiCategory}
     height={500}
     width={500}
-    series={<RadialAreaSeries area={null} type="grouped" />}
+    series={<RadialAreaSeries area={null} type="grouped" interpolation="linear" />}
     axis={<RadialAxis type="category" />}
+    startAngle={-0.5 * Math.PI}
+    endAngle={0.5 * Math.PI}
+  />
+);
+
+export const Custom = () => (
+  <RadialAreaChart
+    height={500}
+    width={500}
+    data={medDateData}
+    innerRadius={0.1}
+    series={
+      <RadialAreaSeries
+        area={null}
+        colorScheme="cybertron"
+        animated
+        interpolation="smooth"
+      />
+    }
+    axis={
+      <RadialAxis
+        arcs={<RadialAxisArcSeries count={5} />}
+        ticks={
+          <RadialAxisTickSeries
+            count={5}
+            tick={
+              <RadialAxisTick
+                line={<RadialAxisTickLine position="outside" />}
+                label={<RadialAxisTickLabel autoRotate />}
+              />
+            }
+          />
+        }
+      />
+    }
+    startAngle={-0.25 * Math.PI}
+    endAngle={0.25 * Math.PI}
+    isClosedCurve={false}
   />
 );
